refactor: reset, theme 파일 경로 변경 #52
